api/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
openai==1.12.0
python-dotenv==1.0.1